fix eclipse warnings and update formatting
git-svn-id: https://svn.apache.org/repos/asf/pivot/branches/2.0.x@1792484 13f79535-47bb-0310-9956-ffa450edef68

+++ core/test/org/apache/pivot/serialization/test/StringSerializerTest.java
@@ -50,7 +50,7 @@ public class StringSerializerTest {
         for (int i = 0; i < b.length; i++) {
             if (i > 0)
                 buf.append(',');
-            int ib = ((int)b[i]) & 0xFF;
+            int ib = b[i] & 0xFF;
             String hex = Integer.toHexString(ib).toUpperCase();
             if (hex.length() < 2)
                 buf.append('0');
@@ -84,7 +84,7 @@ public class StringSerializerTest {
     @Test
     public void writeValues() throws IOException, SerializationException {
         log("writeValues()");
-log("test string = \"" + testString + "\"");
+        log("test string = \"" + testString + "\"");
         // Note: assume the default Charset for StringSerializer is UTF-8, which we are using here
         Serializer<String> serializer = new StringSerializer();
